Extended column testing

+++ src/test/kotlin/fixtures/User.kt
@@ -4,7 +4,7 @@ data class User(
     val id: String,
     val name: String,
     val email: String,
-    val height: Double,
+    val height: Float,
     val age: Int,
     val active: Boolean,
     val address: Address,
